refactor: hide piring package
This package is very low-level and because of that it is not very reusable. Hide it. At least for now, until someone will ask for it. Less public API means less chance to introduce a breaking change.
